Employee API

+++ InsuranceManagementAPI/Data/Repository/IEmployeeRepository.cs
@@ -6,8 +6,8 @@ namespace InsuranceManagementAPI.Data.Repository
     {
         Task<IEnumerable<EmployeeDto>> GetAll();
         Task<IEnumerable<EmployeeDto>> GetAllByName(String EmployeeName);
-        Task<EmployeeDto> GetByID(long id);
-        Task<long> Add(EmployeeDto employeeDto);
+        Task<EmployeeDto> GetByID(int id);
+        Task<int> Add(EmployeeDto employeeDto);
         Task<bool> Update(EmployeeDto employee);
     }
 }
